.history/requirements_20180915104952.txt
kafka-python
conda
python
